Rewrite doImport to use importlib rather than __import__
Python docs really don't want you to use __import__ and strongly
prefer importlib. I think this implementation is arguably easier
to understand and has a key advantage of not hiding ImportError
from the caller (this is important when loading plugins for
formatters and the like since you want to know why your
formatter failed to import).