Enable !include directive in YAML files
A YAML file can now have:

datastore: !include datastore.yaml

to pull in a shared definition into a butler configuration file.
This helps to simplify config duplication in the test configs
such that butlers can be tested with different datastores but
identical storage classes and registries.

To enable this to work a small Loader class is defined
to recognize !include, safe_load is replaced with load (since
custom directives are not considered "safe") and a check is
made to allow nested component keys so that a datastore config,
say, can be standalone and include "datastore:" but also
the butler entry can say "datastore:". Ordinarily this results
in "datastore.datastore" so we strip that out the hierarchy.