Config: Do not split the supplied name if it exists directly
This allows a key of "calexp.wcs" to be found at the top level
if the YAML file defined such a key. This is a bit of a band
aid over the general solution of having to try combinations
all the way down.  We may have to reconsider allowing
compound component names in keys of config files.