Factor out delimiter splitting code and set class value
The delimiter is now set as Config.D in a single location.
Furthermore, the code to process the string supplied by
the caller when addressing a config hierarchy is now
in a single location.  The logic has been expanded
to allow an iterable such as a list as well as a string.
Additionally, if the first character of the string is
not an alphanumeric character then that character
is assumed to be the delimiter for the given string.