Improve Schema support for optional views and limited schemas.
It's now possible to materialize a view as a table by adding
"materialize: false" instead of just removing the "sql" entry;
that might be permit future SchemaBuilders to e.g. add triggers
to populate the table via a SELECT INTO query, for example.

We also now check whether a foreign key target is a view or
is missing due to the limited option before adding it; this
means you don't have to remove or add foreign key entries
whenever you change materialized or limited options.  As a
nice bonus, that means we don't need a limited option for
foreign key entries themselves anymore.