Stop using integer keys in config tests
They do actually work... until `names()` is used and converts
everything to a string which can no longer be found. It's
not worth trying to check for integers as keys as a special
case since it should never happen when reading from a config
file. Therefore remove the test code that was exploiting it.