Implement delayed instantiation of StorageClass (DM-15887)
DatasetType handling of storageClass property was re-implemented to
allow instance construction from StorageClass name. In that case
`storageClass` property uses delayed instantiation (which should only be
done when butler is properly configured). This allows DatasetType use in
a context when butler is not initialised and only name of the
StorageClass is provided (e.g. in Pipeline construction).

Comparison and hash methods of DatasetType class have to be re-implemented
because internal state became more complicated.