A configuration and logic necessary to handle raw in gen2convert.
Gen2 configuration lists 'raw' as DecoratedImage, but CameraMapper's
std_raw transforms it into an Exposure.  In Gen3, it's read now
in directly as an Exposure by the Formatter (if the Formatter uses
DecoratedImage internally, that's the Formatter's business), but we
need to make sure the StorageClass is registered as Exposure.