Standardize APIs for accessing dimension links.
DimensionElement, DimensionSet, and DimensionGraph now all have a
'links' method instead of the mishmash of differently-named properties
they had before.

This also mostly removes the term "primary key" from the documentation
about links; I've left it in a couple of places because it still seems
like the best way to describe the 'link' concept to a complete
newcomer to the codebase.