Implement implied flexibility in DataId construction.
Passing **kwds with new link values along with an existing DataId
now creates a new DataId with those link values changed, rather
than ignoring those kwds.

Unfortunately, this is the latest in a sequence of changes that have
together made DataId construction and expansion *quite* complicated
and hard to follow.  I don't think a refactor is appropriate for this
ticket (it's at least part of DM-17023), but it's now definitely
overdue.