Drop row-wide data ID expansion from selectDimensions.
We only need the per-DatasetType data IDs expanded, because no one
ever looks at the extra metadata associated with the row-wide one.

In the future we should investigate whether we can get rid of the
row-wide data ID entirely, but it does currently appear to be used in
pipe.base.GraphBuilder.