Use transactions in gen2convert.
This is necessary to avoid poor performance on shared filesystems with
SQLite's now-aggressive file locking (by making us lock the DB once
instead of repeatedly locking and unlocking).

It also seems desirable, in the absence of self-consistent
finer-grained transactions on what is inserted into the converted repo
(which we certainly did not have before).