Preserve Python API reference heading hierarchy
There can only be one "Python API reference" section in a module
homepage, and all Python APIs need to be in it. This commit accommodates
the subsectioning within the Python API reference section. Most likely
those subsection will need to be turned into decorative rubric elements
so that the heading typography isn't but off balance; but that's for
another day.