Add property to DimensionJoin to indicate when it should be used.
This somewhat-generalizes the case of VisitDetectorRegion, which
should only be included in a query when the region it provides is
actually needed.  The current query logic identifies this case based
on it being the only DimensionJoin that is a region holder, but I'd
like to start being explicit about that.