Refactor name lookup for templates/formatters/composites
Now wherever a config file has keys that can be StorageClass
names, DatasetType names or "instrument<xxx>" overrides,
the parsing of this is done in one place and a new LookupKey
object is created and used as the key in the mappings.

This provides the ground work for supporting Dimensions
in config files.