Introduce methods to close butlers and registries
Some registries and or butlers that hold registries
may need to do actions to properly close resources.
This adds a close method to call on butlers, which
also calls close on the registry the butler holds.
A specialization for the SqlRegistry is added to
close the connection to the database.

Fixup for shutdown (close) reword