Use <butlerRoot> rather than <root>
Decide against using '{root}' because the curly brace is
special syntax in YAML and would require that the root be
quoted every time.

This change cleans up a test so that the root tag is
retrieved from code and cleans up the documentation
so that the root tag is documented in sphinx without
hard coding the value of the special tag into a
docstring.