Remove need for explicit close method
Nullpools directly give control of connection objects to the code
which calls the connection object. These connections are then closed
when they are returned to the pool. However, like all pools the
connections are wrapped in finalize_fairy methods such that when the
pool is closed this function attempts to clean up resources. Because
the connections is closed at the time it is returned to the pool this
cleanup attempts to use a connection that is already closed, which
throws an exception.

Instead for oracle connections the butler now uses the standard
QueuePool, limited to one connection. This allows the connections to
properly be handled and cleaned up by the pool.

Because the connections no longer need closed, there is no longer a
need for additional close methods in the butler and registry. These
methods are deleted in this commit.