Modified config db entry to be a dict
To allow upcoming changes to specify other DB information separate from
the connection url (e.g., Oracle schema), the db entry in the config was
changed from a single string to a dict of values.  Existing strings
in configs and code was moved to url string inside of db (i.e., .db.url).