Add changing of default schema and setting module
By default, Oracle will use the user's own schema.  Operators need
the ability to write to a different schema.  And until multi-Registry
feature is implemented, users also need to directly access a different
schema.  Added schema to the db dict in the butler config and added
code to change the schema to be used by SQLAlchemy/Oracle.

Also, added code to set the DBMS_APPLICATION_INFO module value to
the executable name providing helpful tracking information to the db
admins.