Make all table names lowercase (with underscores).
This also makes the conceptual names for all concrete dimensions
lowercase, and hence the same as their link names.  The duplication
between link name and dimension name will be removed on DM-17023.
The conceptual names for other tables remain CamelCase, to match the
names of the corresponding Python classes.

In addition to the name changes, this commit also:
 - adds checks that all schema tables and columns are lowercase;
 - removes the Schema.datasetTable attribute.
While it may be a good idea to have named attributes for all major
tables, we shouldn't have one for just this one (and it was only used
inconsistently already).