Rewrite Location and LocationFactory to be more correct
Previously we were a bit loose with our definition of
URI.  This rewrite changes the API to Location such that
it now takes a URI ParseResult as the root of the datastore
and a relative path reflecting the position in the datastore.
This means that we no longer have to deal with strange
relative path URIs that are non-standard for files.

This change removes fromUri from the LocationFactory
API since it makes no sense to define a root URI for the
datastore and then allow another URI to be provided.
No butler code called fromUri.
It also changes the Location constructor but
only LocationFactory was calling it.

There are some shenanigans with trying to deal
with the difference between a posixpath and an os.path
but given that I am testing this on a posix system
there might be some inconsistencies still.

Tests have been added.