Refactor Registry schema system.
This defines Python classes that represent (and document) the YAML
files, allowing some elements of the schema to be generated
programmatically (not yet used, but will be in the future).

It also addresses some style/consistency/normalization issues in the
YAML structure itself (e.g. primary_key->primaryKey, given that we
already had foreignKeys).