Remove Datastore test for bad data IDs.
This test was trying to test that Datastore validation can catch a
data ID that includes a key that isn't part of the template, but
always relied on having a DatasetType with non-normalized dimensions
to do that work.  Now that DatasetType dimensions are normalized, the
Datastore logic fails to catch the bad data ID, but that's okay - it's
not the responsibility of the Datastore to notice that a DatasetRef's
DatasetType and Data ID are inconsistent.