Adjust order of attributes in DatastoreRecord
For datastores the posix_datastore_record database table
had columns whose order was determined by the order of the
entries in the dict defining the data types associated
with entries in the NamedTuple.  I had expected them
to be defined based on the NamedTuple definition so wrote
the Dataclass to match.  In the new implementation the
types definitions and the field orders match because
they are all backed by the dataclass, but in order for
the branch to be compatible with previous gen3 installations
we change the order here to match what was going on
previously.