Remove factory system for Instrument construction.
This was always problematic because it required the right instrument
class to have already been imported in order for it to be registered.
Code that previously used it (mostly in obs_base) now just asks for
the fully-qualified Instrument class name and calls
lsst.utils.doImport, as we do for Formatters and most other plugins.