Allow username or pasword to be None.
Instead of parsing through the file twice with `exists` and then
`getAuth` re-write the `DbAuth.getAuth` method to allow username
and password to be None.

This requires silencing any errors when credentials are
miss-configured from ConnectionStringFactory's side. Short of
having an explicit key in config stating whether a credentials
lookup is desired there is no clear cut way to say if we must
insist that username/password must exist or not.

Instead, we believe the provided db string and try to return it
unchanged as often as possible unless everything is configured
correctly.