Never try to remove the directory created by datastore.put
Removing it can cause problems for concurrent processes trying
to ingest and already believing the directory has been created.
It is safer to let the directory persist even if a problem
occurred with the put.

This also removes the confusion of a log message stating on
error that the directory can not be removed since it contains
files.