Remove the Run class and Registry methods operating on it.
Runs are now always just represented by their string names, the fact
that they have an integer ID is now an implementation detail of
SqlRegistry.

One consequence of this is that "run" and "collection" are now
equivalent in filename templates - they both resolve to the string
name, and there's now no way to use the integer ID instead.  I don't
think we'll miss that in practice.

This leaves us with no way to set the timestamp and host values in
the Registry database, but we weren't using the functionality we had
to do that before, and I don't want to write interfaces without
understanding the use cases for that better.