Implement opaque table operations on new Registry.
These implementations were nominally copied from SqlRegistry, but
adapting them to use Database made them so trivial that we might as
consider them complete reimplementations.

This includes the first tests for the new Registry class, as this is
the first point at which we can do much more than instantiate one.
These are totally new tests, and while we're running them only on
SQLite now, we've set things up so we can run them on other Database
engines easily later.