Transfer testRegions from test_sqlRegistry.py to test_sqlite.py.
This maintains the test coverage we had for this functionality before,
by testing in an in-memory SQLite Registry.  Unfortunately the test
regions used here are enormous, and that makes the test prohibitively
slow on any database that has to do disk I/O.  Fixing this now would be
a distraction from higher-priority work.