Remove core/schema.py and all references to it.
This has now been completely replaced by the new versions in ddl.py.

That drops:

 - support for generating SQLAlchemy object directly from the *Spec
   objects (that's now the responsibility of Database);

 - SQL views (we just construct the equivalent queries on the fly,
   giving us less code to maintain, as views were not supported well
   by SQLAlchemy);

 - SchemaConfig and schema.yaml (specifications are now defined in the
   code that uses them, reflecting their tight coupling).

It also changes region fields to be based on string fields - which have
to be defined with a maximum length - rather than blobs.  We still need
to work out what a reasonable maximum length ought to be for these.