Use new helper classes for opaque tables in Registry.
Right now, this looks like a lot of extra indirection for no change in
functionality, and that's partially correct, because we may never
actually write any other implementations for the opaque table helper
interfaces.  But it sets up a pattern that we'll use in the future for
other (more complicated) Registry functionality where we'll definitely
want multiple implementations.