Move dimensions/storage.py to registry subpackage.
This is *only* a code move, but it splits up storage.py into per-class
files, and that means even more changes to import statements.

The unused ChainedDimensionRecordStorage class has been dropped instead
of moved, as it was originally developed mostly as a thought experiment
for multi-layer Registries and has been superseded in that role by the
prototype page in confluence.