Split and clean up implementations for database-backed dimensions.
This adds a number of checks when constructing
QueryDimensionRecordStorage, which we can really only do under very
restrictive conditions.  That made it clear that it doesn't actually
need to access the database in fetch (so there's no point to
ever caching dimensions that use it).

This change also fixes some problems in querying for dimensions
backed by queries, which we missed before because those same
restrictive conditions mean it's very rare for QueryBuilder to
need to try to join those subqueries into a larger query; a
new unit test addresses this.