Make most Butler attributes private, document the rest.
The registry attribute definitely needs to be public.  The datastore
might for particular datastore implementations (though maybe not any of
the ones we have now).  Other attributes are definitely implementation
details - I was actually thinking about removing config, since we don't
*need* it after construction, and while that looks inconvenient now,
we should leave the possibility open.

This means we have a bit of test code using a private attribute
(._config), but I think a bit of fragility there is better than
removing those tests.