Simplify and clean up DeferredDatasetHandle.
We now resolve the DatasetRef by looking it up in the Registry before
passing it to the proxy object, which means that proxy doesn't need to
hold nearly as much.

This also removes the closure-based hiding of the butler, as we
actually do have valid use cases for getting the butler from these
objects.