Partially mitigate unsoundness in dataset deletion.
This replaces Butler.delete with Butler.prune, which is vectorized and
doesn't "unstore" from the Datastore by default.  That's both because
we don't want to make it too easy to delete things in collections the
current butler doesn't know about, and because PosixDatastore.delete
isn't exception safe, and that makes it way too easy to get a corrupted
repository.