Change how spatial/temporal dimension elements are represented.
A spatial or temporal dimension element now holds a reference to
either self or another element to indicate that it is spatial and/or
temporal.  A reference to another element means that the other element
is a more precise region/timespan that should be preferred, while a
reference to self means it is the most precise region/timespan.

This removes some complex functionality in the query and DimensionGraph
logic that was intended to allow a less-precise region to be used when
explicitly requested.  That was intended to support the jointcal use
case detailed on DM-21904, but we're now solving that a different way.