Fix DimensionGraph's primary key traversal order.
This fixes the logic to correctly handle the case where one
dimension's required dependency is another dimension's implied
dependency.  It also moves it from a private attribute computed at
construction to a construct-on-first-use public property.  It probably
should have always been public (it's used exclusively by Registry),
and it now needs to be construct-on-first-use to avoid cycles during
DimensionUniverse construction.