Implement parsing of time strings in user expression (DM-22173)
Time literal was added to a parser in the form `T'string'`, string can
be one of the supported formats for astropy Time class plus optional
format and scale. Some parsing is done in the parser code to determine
string format and convert it to floating number if needed. For
string-based representations (iso, yday, etc.) it relies on astropy
parsing.