Add DimensionElement.required to more precisely report relationships.
For all current dimension elements in the configuration, we can
assume that element.graph.required is exactly the set of dimensions
used to form the primary key for that element's table.  That's not
going to be true for one we're about to add, so this commit
generalizes the system by adding element.required that can (rarely)
differ.

I'm not thrilled with the naming or the plethora of subtly tricky
distinctions in play here, but doing something about it would be a
much bigger change than I want to make in the near future; I already
experimented with a few alternatives that quickly got out of hand and
had to be reverted.  The name of the new attribute at least has a nice
consistency and symmetry with other names in the system that should
steer the user in the right direction _most_ of the time, and the
subtle distinctions are now documented.