ButlerURI can discern dir-like from file URIs.
When creating Keys in AWS S3 only keys ending with '/' will
appear and behave as directories. Thus, there is a difference
between inserting `/abs/dir` and `/abs/dir/` keys into a Bucket.
This affects creation of Datastore roots and Butler.makeRepo
functionality which will then create non-intuitive paths in S3.

Added dirLike attribute to ButlerURI. Attribute tracks whether
the ButlerURI ends with a separator or not.
Added forceDirectory argument to ButlerURI. When set to True it
will force a trailing separator to exist.
Forced LocationFactory, Butler.makeRepo and Location datastoreRootUri
argument to always forceDirectory. These are places where we *know*
that the given string *must* be a directory.

Rearranged the Location testButlerUri cases because they were hard
to keep track of.

Added explicit additional check on whether a Bucket exists in
Butler.makeRepo. Now raising a ValueError that the bucket doesn't
exist.

Moved setting of Config.configFile to __initFromFile from
__initFromYamlFile.