Minor cleanups for Registry.queryDatasets.
The len(collections) call that was removed here was a bug: it could
be testing the length of a string collection name rather than the
actual number of collections.  The addition of code to standardize
the collections expression at the top of queryDatasets doesn't quite
fix that, because what we actually care about is the number of
recursively-expanded child collections when CHAINED collections are in
play, so the safe thing to do is just to remove that check.

But I've also kept the code to standardize the collections anyway - it
means we only do that once when queryDatasets self-recurses to handle
different dataset types.