Make note about warning message from test
InMemoryDatastore can not ingest so in a ChainedDatastore
it will complain when a dataset is trashed that it knew
nothing about. We may want to have finer control of logging
when a ChainedDatastore is involved since chaining by definition
allows datasets to be added to some child datastores but
not others.