Improve connection string matching by ignoring driver value.
Matching is only performed on the dialect part of the
connection string scheme and are otherwise ignored in
the context of authentication.
Driver value in the connection string from registry
config is preserved.

Added tests for Oracle with and without Oracle
Wallet authentication to ensure ConnectionStringFactory
correctly ignores not-standard DB URLs.

Updated tests to match the changes.