Replace DataCoordinate stringification with something more useful.
We now just implement __repr__ and let __str__ delegate to that, and
aim to make it informative and non-confusing because we've never been
able to make the string eval'able as in the ideal case.

Note that the biggest problem with the old repr (which is what is used
by __str__ in built-in containers of data IDs, regardless of what we'd
prefer) is that it printed the full DimensionGraph, which often has
implied dimensions and hence more "keys" than the mapping interface
actually provided access to.  That it impossible to infer what the
dict-like form actually was, and it's the dict-like form that people
actually want to know.