Remove ill-conceived data ID consistency checking.
The only code that used this was in QuantumGraph generation in
pipe_base, and that's already been replaced to improve performance.

More importantly, the parts of this that involved spatial and temporal
consistency never made sense - we have clear use cases where the naive
intersection relationships the butler uses by default for these aren't
appropriate (e.g. jointcal), and I should have realized earlier that
this code didn't leave room for them.