Add initial implementation of attribute manager (DM-25354)
The interface for the manager defines few trivial operations, additional
methods will be added if needed as we discover new use cases. An
important part of this ticket is to define schema for attributes table
that should stay stable for long time. For this reason I decided to use a
simplest possible key-value storage with both key and value being
strings, reasonably long key (1kB) and value (64kB). I think this should
be sufficient for our current use cases.