adopt log implementation from pipetask
* change log-level to optionally accept key-value pairs
* configure python and C++ logging
* init (add handler) to the log once
* log level can be set on modules or globally at any time
* add a logging unit test

Changes the type passed from the log-level option log_level
to be a dict, any non-kv value is put in the dict in an
entry whose key is an empty string. If any keys are
repeated (including no key, in which case the key will be
an empty string) then the last entry takes precedence.

Applies the Choice (accepted input values) to the value;
any key will be accepted.

add logging unit test