Move where parameter validation occurs
Parameters for read-only components are problematic since it
is not entirely clear whether the parameters should be applied
to the component that is used to calculate the read-only
component, or should be applied to the calculation of
the read-only component itself.

The complication is that assemblers must support the same
parameters as formatters (otherwise in-memory datastore
can not function) and also can apply parameters that
the formatter did not understand. In the current implementation
this means that the assembler can only see the final storage
class and at that point parameters might not be relevant.

As concrete examples. In the test suite I have added
a "counter" read-only component that returns the
number of elements in metrics.data.  This means that
the assembler sees an Integer storage class and integers
aren't amenable to parameters. Instead the "slice" parameter
is assumed to apply to the "data" component and then
"counter" is calculated on the sliced "data".

Another example is bbox in Exposure. This can be applied
to the "image" component. Where should parameters go?
To the calculation of the "image" or to the calculation
of the bounding box (origin is the only one). Once
the bounding box is created though the "origin" parameter
has no meaning so "origin" is only relevant when passed
to the "image" and can't be relevant to the assembler.

Therefore read-only components for disassembled
composites do not receive any parameters.