Move mypy execution to root directory, and run it on a directory.
This is the more common and better-supported way to run mypy, but it
was previously blocked by the fact that our root (lsst) package was
a namespace package.  This change should make no external difference in
what mypy does right now, but I suspect it will be easier to maintain
and I can already say it's much more compatible with at least some
in-editor mypy integration tooling.