Add more transfer modes to posix export
We really need to generalize the code that transfers one
file to another location so that ingest and export use
the same code and preferably can be shared amongst different
datastores. You can imagine ingest from S3 to local file or
export from S3 to local file or export from local file
to S3.  It is not scalable to put this code in multiple
places in multiple datastores.