Make SAVEPOINT usage opt-in in transactions.
This is the behavior I originally wanted for Database when that class
was first written, but Andy Salnikov correctly pointed out that the
instance-attribute approach I used to track the current transaction
state was unsafe.  I think this approach - which uses the free-form
'info' attribute of SQLAlchemy's Connection object - solves that
problem.