Use local tempdir to work around relsymlink problem
on macOS the /var folder is a link to /private/var which
means that when you create a temp directory in a test
and put a file in it, once you readpath you end up
in a completely different location that is nowhere
near the test file that you are trying to use
with relsymlink.  Using a local tempdir fixes
this anomaly.