Fix Timespan behavior and make it half-open.
Trying to make this half-open (and test it against PostgreSQL
operators, WIP for another commit) revealed some fundamental problems.

We want this to be half-open because none of our existing use cases
care about whether bounds are open or closed, but calibration
collections will want to guarantee that the validity ranges for a
particular dataset type and data ID are exclusive, and that's _much_
simpler if adjacent ranges do not overlap.