Rework DimensionRecord construction and fields.
This makes a lot of code a lot safer, as it no longer relies on a
largely unenforceable convention about the ordering of fields.

It also makes `DimensionRecord.fromDict` unnecessary, and hence
that's being removed.

Finally, while this spreads out the logic that assembles and
disassembles Timespan objects and overall makes it messier, this is
all in preparation for abstracting over the database representation of
Timespans on a future commit, and it will make more sense then.