Remove check for artifact existence on datastore.put
If registry doesn't think it should exist then
datastore should allow the overwrite.

This can go wrong if a file template is not unique such
that different datasets can end up with the same file name.

It may also be possible for the file to exist but be
in the trash table. In that situation the file might get
deleted when the trash is emptied.