use static methods in shared arguments
Even though arguments don't use a common abstract base class
(yet...) they need to follow the change made to shared options,
to use a static method for 'defualtHelp' instead of a class
property, because the api needs to match.

There's probably room for an MWArgumentDecorator class here that
would be very similar to MWOptionDecorator. They could even have
a common MWParameterDecorator base. TBD when it's just insanity.
For now it seems Good Enough to have the option decorator, and
just make this change; arguments are infrequent enough.