adds trial "forward option" feature
Option forwarding allows CLI option values passed to one subcommand
to be "forwarded" to the next subcommans without the user having
to re-enter the same options for the next subcommands.

For example, "pieptask qgraph <butler options> run" allows butler
options passed to the qgraph subcommand to be used by the run
subcommand.