Restructure dataset deduplication queries.
This moves the window function / partition stuff to the outermost query
rather than a subquery.  This will be necessary for deduplication of
CALIBRATION collection lookups, and it should enable deduplication of
dataset queries that have more dimensions than those needed to identify
the dataset.

It's one more straw on the back of the creaking Query/QueryBuilder
design camel - the way we use these classes is now pretty different
from how they were originally conceived, and there has been a steady
increase in tight coupling between classes, conditionals, and other
code smells.  It's in need of a refactor, but this isn't the ticket
for it.