Remove DimensionGraph inheritance from DimensionUniverse.
These was no need for DimensionUniverse to behave like a DimensionGraph
in any existing context, and in fact the only useful functionality it
inherited from DimensionGraph (some sort-order indices) were not used
by any non-universe DimensionGraph instances.

Breaking this inheritance opens the door for some dimensions to not
need to be fully declared, and hence the universe of all known
dimensions to be open in a sense.  That is hinted at here by the names
of the new DimensionUniverse methods that replace public access to the
elements and dimensions - those may someday provide access to the
"static" part of the universe, which won't be the whole thing.