reduce shared option boilerplate code
MWOptionDecorator used to be a subclass for shared option
decorator types. This changes it to be a factory that
generates a callable type that returns a click.option
with the shared parameters prepopulated (using
functools.partial). It also implements a few helper
functions for e.g. getting the name of the argument
with which the option will call the command function.