Move repository import/export code into separate package.
This should not change functionality at all; it's just moving code
around.  A separate package gives the code a bit more room for
expansion and more clearly separates between interfaces and
implementations.

Moving it out of 'core' fixes the existing circular dependency issues:
the import/export code needs to depend on both Registry and Datastore,
and hence shouldn't go in core.  But the registry tests depend on it,
so there is one new circular dependency in registry.tests, but because
tests conceptually depend on everything I'm not bothered by this
(maybe we should have named that 'registry_tests' instead of making it
a registry subpackage, but I'm not going to).