Store dimensions configuration in registry database (DM-26407)
Dimensions are now stored in registry attributes table with a key
"config:dimensions.yaml", serialized as YAML data. This change means
that we cannot use SQLite in-memory datqabase for tests that make
multiple connections (multiple Butler instances) so I had to update many
unit tests to use on-disk database instead. With this commit dimensions
are still part of full butler config but they will be separated in next
commits.