Add governor dimension restrictions to collection chains.
Instead of just a DatasetTypeRestriction, collection chain entries
are now associated with a CollectionContentRestriction (new), which
combines that DatasetTypeRestriction with a
GovernorDimensionRestriction (also new).

This is missing a good way to set up those new restrictions via some
kind of user-friendly expression; there are currently only "advanced
usage" APIs for that.  But that may actually be all we need.

This is a backwards incompatible SQL schema change.  The YAML export
format has changed as well, but there is backwards compatibility
logic in the reader.