improve 'prune-collections' arg error reporting
Adds an execption type (inherits from TypeError) for the ways
arguments to Butler.pruneCollections can fail. This allows the
CLI script function to catch the exception, know the reason for
the failure, and format a message to report the error in a way
that will make sense on the command line.
If this seems like an acceptible solution to reviewers this pattern
may be used elsewhere to improve CLI error reporting as needed.