Check consistency and completeness of governor dimensions in queries.
This restructures the QueryWhereExpression into a pair of classes,
with the new QuereWhereClause holding information about what's in the
query and picking up some attributes that were previously part of
QuerySummary.  QuerySummary now has a QueryWhereClause instead of a
QueryWhereExpression, and transforming from the latter to the former
(via QueryWhereExpression.attach) is what performs all of the new
checking.

That checking is implemented in registry/queries/expressions.py, which
includes some small modifications to InspectionVisitor to make it
something the new CheckVisitor can delegate to.  CheckVisitor and its
Summary helper classes is where all of the new checks happen.

Because checking can produce some false positives, all of the Registry
query methods now have a 'check' kwarg that can be used to turn it
off.  At present I'm not planning to expose that in command-line
utilities - I'm hoping we can get by with always checking those
(especially the queries that come from QuantumGraph generation).

This implements DM-26950, but I'm reserving the right to follow this
with a commit that changes the default behavior to check=False, and
let DM-26590 to revert that commit later, if check=True causes too much
disruption downstream.