Add intermediate dimension storage ABCs and rework construction.
DimensionElement is now a factory for its record storage class, via
a fully-qualified class name in configuration.

This is *maybe* a step backward in terms of dependencies - it gives
DimensionElement a (weak) dependency on classes we otherwise consider
"downstream" in the registry subpackage, but it's a big step forward
making those record storage classes more fully responsible for things
they should have always been responsible for, like caching, and for
avoiding a lot of smelly isinstance checks in record storage in the
future as the database representations of different kinds of dimensions
diverge further.  It also gives a lot more flexibility in configuring
dimension storage, because we can pass arbitrary kwargs directly from
config.

This transformation also doesn't feel complete: the viewOf and hasTable
properties of DimensionElement really belong on the record storage
classes instead (and the new implementation of
DatabaseDimensionElement.viewOf reflects that by peeking at what are
supposed to be - from its perspective - opaque keyword arguments to
forward).  But fixing that is best put on another commit (and maybe not
done now), as this commit is plenty big already.