Rewrite tests for spatial overlap queries.
We used to have two tests - one that just tested very little (i.e. that
we got no overlaps when there are no regions) and one that tested a
small number of enormous (significant fractions of the sphere) regions
and could only be run in-memory SQLite without bogging down.

We now have one more complete test that runs on all databases and uses
realistic data from an HSC RC2 dump that was done a while ago.