Remove content restrictions from collection search expressions.
This also removes those restrictions from CHAINED collection
definitions.

The *Restriction classes in wildcards.py are still present, as they may
still be useful in the future.  "May be useful" is not a good reason
to keep them around, given that they'll always be in the git history,
but I'll remove them on a separate commit.