Cleanup the core test that loops over multiple butler.put variants
Previously if an earlier test in the loop failed by definition
all tests would fail after that because a dataset would
exist that could not be replaced.  Now each time round the
loop a new tag/collection is used so even if one fails
the next attempts (and the attempt outside the loop)
will not be affected.