allow split_kv CLI util to preserve default value
In some cases, key-value options want to add to the default value
instead of overwriting it entirely. An example is --log-level,
where the option default may be set:
@log_level_option(default=["WARNING"])
and the caller may want to pass in a module level:
--log-level lsst.daf.butler=DEBUG
But the overall WARNING level should not be overwritten for the
option. This adds an argument to split_kv called add_to_default
that supports this mode when True.

This removes the check for duplicate values (previously we would
raise if duplicate values were encountered). There was a note also
that suggested we may want to accumulate values, so keys are passed
more than once then the value would be a list. This was not
implemented, but may want to do that at some point.