Remove tags and chains arguments from Butler constructor.
This slightly complicates a couple of CLI tests, but it removes a lot
more unnecessary complexity from Butler itself (and the tests for this
functionality); this was convenience functionality that just doesn't
add much value.  It was more useful when we were first adding CHAINED
collections and transitioning from using TAGGED collections in
processing runs.