Rewrite content summary classes and use them to simplify query checks.
DatasetTypeRestriction is no longer needed at all, and is now gone.

GovernorDimensionRestriction was included in other code only because
it will be used in the future, but now it's useful enough to simplify
the query expression checking code that was doing very similar things.

CollectionSummary (previously CollectionContentRestriction) hasn't
ever been used, but I plan to soon.