Move special-casing of '*' and empty lists into globToRegex.
All of the command-line scripts that used globToRegex previously had
to special-case the empty list it would return when '*' or an empty
input list was encountered.  But we already had a consistent sentinal
value - `...` - that we used for all of these cases.  By making
globToRegex return that when '*' or an empty list is encountered, we
can simplify the scripts and fix DM-28527, which was due to the lack
of this special-casing in queryCollections.