Avoid long-lived connections in Database (DM-26302)
Database connections are now short-lived by default, they are created
either for the duration of a query or transaction. Additionally `session()`
context can be used for persistent connection that does not start
transaction, this is needed to manage lifetime of temporary tables.
Re-enabled connection pool use in database engines, this will need special
care in fork-based mutiprocessing.