Add separate registry method for importing datasets.
New `_importDatasets` method is separate from `insertDatasets` to make it
more explicit what is the purpose of each. I have added unit tests for
UUID manager and a YAML file with UUID IDs instead of integers. Also
added is a unit test which imports the same datasets twice (for UUID
case only).