Allow trash to take iterable of dataset refs
This is much much faster than calling trash on each ref
separately. Also remove the existence check since if we are
asking for a dataset to be trashed it is not really that
important if the dataset has already disappeared and that
could just be caused by someone else prunning a collection
in parallel.